Licenses and ordering refactor
